refactor: extract schedule modal logic into useSchedule
